chore: Add initial Micronaut project (#1)
This commit adds the initial Micronaut project and basic scripts.

The `Dockerfile` provided out-of-the-box by Micronaut was modified
to leverage multistage Dockerbuild and cut down the size of the
release image. The JDK/JRE was also changed from 13 to 11 to put
project in sync and silence build warnings. An upgrade should be
investigated.

A `test.sh` script is not yet implemented as there are no tests of
note at this time. Ultimately, this script will spin up all
dependencies and execute the full test suite. When `test.sh` is
added, `localbuild.sh` will be modified to only execute the linter
and unit test suite.

Docker operations are handled manually at this time to gain
experience with Docker. Using the jib-gradle-plugin resulted in a
similar sized image. However, it may be beneficial to leverage the
plugin in the future.