fix: update Kotlin plugin version to fix Gradle build
A conflict arose with version 1.3.50 of the Kapt plugin and a
dependency used in the project that caused the build to fail.

Also worth noting is that there's no proof that is exactly what
occurred, but the build failure occurred seemingly out of the
blue with no code changes or manual dependency version updates.