feat: Add Roaster DB model & In Memory Repository (#3)
This commit adds a first take at the Roaster/Source model.
This model is a work in progress. For now, this data will
very infrequently change and will be supported by an in
memory datastore that is initialized from a JSON file at
startup. This will be refactored to be backed by a different
datastore once this model is more firm.